requirements.txt
accelerate==0.24.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
# autoattack==0.1
babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
braceexpand==0.1.7
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
geotorch==0.3.0
gitdb==4.0.10
gitpython==3.1.31
huggingface-hub==0.14.1
idna==2.10
inflection==0.5.1
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyter-events==0.6.3
jupyter-server==2.10.0
jupyter-server-terminals==0.5.3
jupyterlab-pygments==0.3.0
kiwisolver==1.4.4
lit==16.0.3
markupsafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook-shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.99
open-clip-torch==2.19.0
overrides==7.4.0
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
platformdirs==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pyyaml==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.25.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robustbench==1.1
scikit-learn==1.3.2
scipy==1.10.1
send2trash==1.8.2
sentencepiece==0.1.98
sentry-sdk==1.22.2
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.5.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.3
# torch==2.4.0+cu124  # 单独安装
# torchaudio==2.4.0+cu124  # 单独安装
torchdiffeq==0.2.3
# torchvision==0.19.0+cu124  # 单独安装
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
# transformers==4.30.0.dev0
triton==3.0.0
typing-extensions==4.9.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.15
wandb==0.17.7
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
xxhash==3.2.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2